refactor: Changed +page.server.js to +page.js 🏗️
Changed +page.server.js to +page.js, so that data can also be fetched through the client when navigation is handled clientside by Svelte. This means that the URL can be used to filter through the posters when the search function is used, by simply letting Svelte call the load function on the client.